+++ src/utils.c
@@ -1,9 +1,10 @@
 /* SPDX-FileCopyrightText: 2021 Comcast Cable Communications Management, LLC */
 /* SPDX-License-Identifier: Apache-2.0 */
-
-#include <stddef.h>
 #include <stdint.h>
-#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include "utils.h"
 
 /*----------------------------------------------------------------------------*/
 /*                                   Macros                                   */
@@ -33,10 +34,39 @@
 /*----------------------------------------------------------------------------*/
 /*                             External Functions                             */
 /*----------------------------------------------------------------------------*/
-int main(int argc, char *argv[])
-{
-    (void)argc;
-    (void)argv;
 
-    return 0;
+
+size_t xa_strnlen(const char *s, size_t maxlen)
+{
+    for (size_t i = 0; i < maxlen; i++) {
+        if ('\0' == s[i]) {
+            return i;
+        }
+    }
+
+    return maxlen;
+}
+
+
+char *xa_strdup(const char *s)
+{
+    return xa_strndup(s, SIZE_MAX);
+}
+
+
+char *xa_strndup(const char *s, size_t maxlen)
+{
+    char *rv = NULL;
+
+    if (s && (0 < maxlen)) {
+        size_t len = xa_strnlen(s, maxlen);
+
+        rv = (char *)malloc(len + 1);
+        if (rv) {
+            memcpy(rv, s, len);
+            rv[len] = '\0';
+        }
+    }
+
+    return rv;
 }